Experimenting with querying an Azure Cosmos DB
